+++ src/Model/Actors/IActor.h
@@ -15,11 +15,11 @@ namespace Actors {
         virtual void Hit(const GameCard::Cards &) = 0;
         virtual void SetCard(const GameCard::Cards &) = 0;
         virtual void DoubleDown(const GameCard::Cards &, const GameCard::Cards &) = 0;
-        virtual void GetResult(double) = 0;
+        virtual void GetResult(int) = 0;
         [[nodiscard]] virtual bool BlackJackCheck() const = 0;
 
         [[nodiscard]] virtual const GameCard::Hand & ShowHand() const = 0;
-        [[nodiscard]] virtual double GetPlayerCost() const = 0;
+        [[nodiscard]] virtual int GetPlayerCost() const = 0;
     };
 }
 
@@ -27,18 +27,18 @@ namespace Actors {
     struct OfflinePlayer : public IActor{
     protected:
         GameCard::Hand m_hand;
-        double m_bank;
+        int m_bank;
     public:
-        OfflinePlayer(double bank);
+        OfflinePlayer(int bank);
 
         void Hit(const GameCard::Cards &) override;
         void DoubleDown(const GameCard::Cards &, const GameCard::Cards &) override;
         void SetCard(const GameCard::Cards &) override;
-        void GetResult(double) override;
+        void GetResult(int) override;
         [[nodiscard]] bool BlackJackCheck() const override;
 
         [[nodiscard]] const GameCard::Hand & ShowHand() const override;
-        [[nodiscard]] double GetPlayerCost() const override; // можно создать бд отдельно от player и хранить деньги там, player просто будет ключем, не управляя от класса сколько у него денег и тп. Все операции над деньгами там
+        [[nodiscard]] int GetPlayerCost() const override; // можно создать бд отдельно от player и хранить деньги там, player просто будет ключем, не управляя от класса сколько у него денег и тп. Все операции над деньгами там
     };
 
 }
